Cluster.known_hosts_file is now a property.
When moving a storage file from a directory to another, the path to
the known_hosts file is not updated, unless the cluster attribute is
updated.

However, since we keep the `known_hosts` file together with the
cluster file, it doesn't make much sense to have an independent
variable. Instead, known_hosts_file is now a property, computed
starting from the `storage_path` where the cluster is actually stored.