When running the "resize" command, the "extra" fields were not getting passed down to the new node. For Google Compute Engine instances, this meant that the SSH keys were not getting set up correctly, and then Elasticluster could not connect to the nodes.
I fixed this for both the case where the first node is used as
a template as well as the case where an explicit template is
specified. I moved two lines out of the "for" loop:

    image_user = conf['login']['image_user']
    userdata = conf_kind.get('image_userdata', '')

as best as I can tell, they don't need to be set with each
iteration.