Remove `--verbose` option from subcommands.
It turns out that the `--verbose`/`-v` option was being redefined in
each subcommands' argument parser; *however*, the subcommand option was
being ignored.  In other words, depending on where `--verbose` was found
on the command line, it would affect logging or be ignored:

* This invocation would do the right thing and make ElastiCluster more
  chatty::

       elasticluster --verbose start cluster

* This use of `--verbose` would have been *silently* ignored::

       elasticluster start --verbose cluster

Now the only supported form is the first one; the second gives instead
an error.

One day, I'm going to rewrite the entire CLI and option processing using
"click" instead but that day might not come soon...