Add compatibility symlink for pvfs2, ceph and jenkins.
The mentioned playbooks include `common/tasks/ssh_host_based_authentication.yml`
but due to the different include search path that Ansible uses when
`include:` statements occur in a role task file or in a top-level
playbook, there is no way of specifying an include path that works for
both.

The symlink is a temporary fix; see issue #293 for more information.