SLURM: Use host name to contact master node.
Instead of hard-coding the controller node's IP address in
`slurm.conf`, we just use the host name and let SLURM resolve it using
the normal system means.

In #583 it was pointed out that IP addresses change upon reboot in GCP
and AWS, so we need to make the configurations robust against those
changes.